Fix adding items, implement commands. - Implement !ticker command (does not save to server config, at least for now) - Implement !notice - Simplify !map and !item. - Fix CharacterItems.Add() packet.

+++ src/Maple/Characters/CharacterItems.cs
@@ -109,7 +109,7 @@ namespace Destiny.Maple.Characters
                             .WriteByte(1)
                             .WriteByte((byte)InventoryOperationType.AddItem)
                             .WriteByte((byte)item.Type)
-                            .WriteByte((byte)item.Slot);
+                            .WriteShort(item.Slot);
 
                         item.Encode(oPacket, true);
